Problem: Weird Algorithm - CSES Status: done (fix: just use variable LONG).

+++ src/Main.java
@@ -5,7 +5,7 @@ import java.util.*;
 class Main {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
-        int N = scanner.nextInt();
+        long N = scanner.nextLong();
         System.out.print(N + " ");
 
         while(N != 1){
